requirements.txt
matplotlib >= 3.9.0
numpy >= 2.1.0
tqdm >= 4.66.0
PyYAML >= 6.0.0
